Use literal legacy answer triggers
